bigdata-project/backend/requirements.txt
django-filter==2.3.0
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.10
joblib==0.14.0
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.15.0
sqlparse==0.3.0
urllib3==1.25.10